+++ src/kataloger/data/catalog.py
@@ -2,6 +2,8 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
 
+from kataloger.helpers.backport_helpers import remove_suffix
+
 
 @dataclass(frozen=True)
 class Catalog:
@@ -10,8 +12,7 @@ class Catalog:
 
     @staticmethod
     def from_path(path: Path) -> "Catalog":
-        catalog_name = path.name.removesuffix(".versions.toml") if path.name.endswith(".versions.toml") else path.name
         return Catalog(
-            name=catalog_name,
+            name=remove_suffix(path.name, ".versions.toml"),
             path=path,
         )