+++ changelog.md
@@ -1,3 +1,5 @@
 ### Changelog
 
-* Repositories configuration file replaced with kataloger configuration file, where beside repositories can be specified catalog paths and parameters. 
+* Added support for more library and plugin notations in catalog.
+* Repositories configuration file replaced with kataloger configuration file, where beside repositories can be specified catalog paths and parameters.
+* Minimum required python version lowered to 3.8.